perf: Use ncc
