+++ .gitignore
@@ -72,3 +72,5 @@ $RECYCLE.BIN/
 *.lnk
 
 # End of https://www.toptal.com/developers/gitignore/api/linux,windows,macos
+
+node_modules/
